Create model for comments
Created a model for comments that will allow users to comment on
articles.

This has been nested within articles so that an article can have many
comments but a comment belongs to one article.

A route has also been set up for comments as well and this is nested
within the articles route.